Completed till section 3.7 of assignment

+++ src/AirVehicle.java
@@ -2,7 +2,7 @@ abstract public class AirVehicle extends Vehicle {
     private double maxAltitude;
 
     //constructor
-    AirVehicle(String id, String model, double maxSpeed, double currentMileage, int maxAltitude) {
+    AirVehicle(String id, String model, double maxSpeed, double currentMileage, double maxAltitude) {
         super(id,model,maxSpeed,currentMileage);
         this.maxAltitude = maxAltitude;
     }
